+++ lib/src/widgets/inputs/mp_azimuth_picker_widget.dart
@@ -1,6 +1,7 @@
 import 'dart:math' as math;
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:mapiah/src/auxiliary/mp_interaction_aux.dart';
 import 'package:mapiah/src/auxiliary/mp_numeric_aux.dart';
 import 'package:mapiah/src/constants/mp_constants.dart';
 import 'package:mapiah/src/painters/mp_compass_painter.dart';
@@ -38,7 +39,28 @@ class _MPAzimuthPickerWidgetState extends State<MPAzimuthPickerWidget> {
     widget.azimuthTextController.text = _azimuth.toStringAsFixed(1);
   }
 
-  void _updateAzimuth(double newAzimuth, {bool updateTextField = false}) {
+  @override
+  void didUpdateWidget(covariant MPAzimuthPickerWidget oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    final double normalizedInitialAzimuth = MPNumericAux.normalizeAngle(
+      widget.initialAzimuth,
+    );
+
+    if (!MPNumericAux.nearlyEqual(_azimuth, normalizedInitialAzimuth)) {
+      _updateAzimuth(
+        normalizedInitialAzimuth,
+        updateTextField: true,
+        notifyChange: false,
+      );
+    }
+  }
+
+  void _updateAzimuth(
+    double newAzimuth, {
+    bool updateTextField = false,
+    bool notifyChange = true,
+  }) {
     newAzimuth = MPNumericAux.normalizeAngle(newAzimuth);
 
     setState(() {
@@ -47,7 +69,10 @@ class _MPAzimuthPickerWidgetState extends State<MPAzimuthPickerWidget> {
         widget.azimuthTextController.text = newAzimuth.toStringAsFixed(1);
       }
     });
-    widget.onChanged(newAzimuth);
+
+    if (notifyChange) {
+      widget.onChanged(newAzimuth);
+    }
   }
 
   void _handleTextInput() {
@@ -139,11 +164,8 @@ class _MPAzimuthPickerWidgetState extends State<MPAzimuthPickerWidget> {
       center: Offset(compassRadius, compassRadius),
       position: localPosition,
     );
-    final Set<LogicalKeyboardKey> logicalKeysPressed =
-        HardwareKeyboard.instance.logicalKeysPressed;
     final bool isCtrlPressed =
-        logicalKeysPressed.contains(LogicalKeyboardKey.controlLeft) ||
-        logicalKeysPressed.contains(LogicalKeyboardKey.controlRight);
+        MPInteractionAux.isCtrlPressed() || MPInteractionAux.isMetaPressed();
     final double adjustedAngle = isCtrlPressed
         ? (angle / mpAzimuthConstraintAngle).round() * mpAzimuthConstraintAngle
         : angle;